Semble regler les problemes de pc negatif et de gestion de l'IDX_MOD, a tester plus en profondeur

+++ vm/src/cw_ins_fork.c
@@ -6,7 +6,7 @@
 /*   By: ademenet <ademenet@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/07/13 15:16:29 by ademenet          #+#    #+#             */
-/*   Updated: 2016/08/09 12:13:52 by ademenet         ###   ########.fr       */
+/*   Updated: 2016/08/12 15:41:21 by ademenet         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -40,13 +40,13 @@ void		cw_ins_fork_duplicate_reg(t_champion *new, t_champion *old)
 
 int			cw_ins_fork(t_proc *proc, t_champion *tmp, t_ocp *ocp)
 {
-	unsigned int	p;
+	short	p;
 
-	p = cw_get_data_dir(proc, tmp, (tmp->pc + 1) % MEM_SIZE, 2);
+	p = (short)cw_get_data_dir(proc, tmp, (tmp->pc + 1) % MEM_SIZE, 2);
 	cw_lst_push(&proc->champions, cw_lst_new(tmp->header, tmp->num));
 	proc->champions->ins = NULL;
 	proc->champions->pc_origin = tmp->pc_origin;
-	proc->champions->pc = ((tmp->pc + p) % IDX_MOD) % MEM_SIZE;
+	proc->champions->pc = (tmp->pc + (unsigned short)(p % IDX_MOD)) % MEM_SIZE;
 	proc->champions->inst_c = 0;
 	proc->champions->lives = 0;
 	proc->champions->id = tmp->id;
